feat!: rename to pretix-signature-capture and refresh metadata
- Rename module pretix_signature_question -> pretix_signature_capture
  (directory + static/ + templates/ subdirs; preserved as renames).
- pyproject.toml: package name, authors, maintainers, homepage, entry
  points, dynamic version path, requires-python>=3.11, keywords,
  description.
- apps.py: PluginConfig.name + verbose_name + PretixPluginMeta name +
  author + compatibility (pretix>=2026.0.0). Bump RuntimeError.
- __version__ -> 2.0.0a1.
- signals.py + presale_head.html: new template/static paths.
- setup.cfg, Makefile, MANIFEST.in, .update-locales.sh, GH workflow
  path globs: updated to new module name.
- README.rst rewritten with new branding, install instructions, and
  migration notes for users of pretix-signature-question[-2].
- planning.md added to track release-2.0 cross-cutting decisions and
  per-issue progress.

Closes #1.
