+++ .github/workflows/tests.yml
@@ -4,11 +4,11 @@ on:
   push:
     branches: [ main, master ]
     paths-ignore:
-      - 'pretix_signature_question/locale/**'
+      - 'pretix_signature_capture/locale/**'
   pull_request:
     branches: [ main, master ]
     paths-ignore:
-      - 'pretix_signature_question/locale/**'
+      - 'pretix_signature_capture/locale/**'
 
 jobs:
   test:
