+++ MANIFEST.in
@@ -1,5 +1,5 @@
-recursive-include pretix_signature_question/static *
-recursive-include pretix_signature_question/templates *
-recursive-include pretix_signature_question/locale *
+recursive-include pretix_signature_capture/static *
+recursive-include pretix_signature_capture/templates *
+recursive-include pretix_signature_capture/locale *
 include LICENSE
 exclude .gitlab-ci.yml
