+++ pretix_signature_capture/locale/de/LC_MESSAGES/django.po
