+++ pretix_signature_capture/locale/de_Informal/LC_MESSAGES/django.po
