+++ pretix_signature_capture/signals.py
@@ -4,8 +4,8 @@ from django.template.loader import get_template
 from pretix.presale.signals import html_head
 
 
-@receiver(html_head, dispatch_uid="pretix_signature_question_html_head")
+@receiver(html_head, dispatch_uid="pretix_signature_capture_html_head")
 def html_head_presale(sender, request=None, **kwargs):
-    template = get_template('pretix_signature_question/presale_head.html')
+    template = get_template('pretix_signature_capture/presale_head.html')
     ctx = {}
     return template.render(ctx)
