+++ pretix_signature_capture/static/pretix_signature_capture/main.js
