--- pretix_signature_question/templates/pretix_signature_question/presale_head.html
@@ -1,8 +1,0 @@
-{% load static %}
-{% load compress %}
-
-{% compress js %}
-    <script src="{% static "pretix_signature_question/jSignature.min.js" %}"></script>
-    <script src="{% static "pretix_signature_question/main.js" %}"></script>
-{% endcompress %}
-
