+++ setup.cfg
@@ -14,13 +14,13 @@ skip = setup.py
 use_parentheses = True
 force_grid_wrap = 0
 line_length = 88
-known_first_party = pretix_signature_question
+known_first_party = pretix_signature_capture
 
 [tool:pytest]
 DJANGO_SETTINGS_MODULE = pretix.testutils.settings
 
 [coverage:run]
-source = pretix_signature_question
+source = pretix_signature_capture
 omit = */migrations/*,*/urls.py,*/tests/*
 
 [coverage:report]
